Extracting code: single type per file
