+++ src/NuGetGallery/Queries/IPackageVersionsQuery.cs
@@ -1,0 +1,15 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace NuGetGallery
+{
+    public interface IPackageVersionsQuery
+    {
+        Task<IEnumerable<string>> Execute(
+            string id,
+            bool? includePrerelease = false);
+    }
+}
