+++ src/NuGetGallery/Queries/PackageVersionsQuery.cs
@@ -6,61 +6,10 @@ using System.Collections.Generic;
 using System.Data.Entity;
 using System.Globalization;
 using System.Linq;
-using System.Net.Http;
 using System.Threading.Tasks;
-using Newtonsoft.Json.Linq;
-using NuGet.Services.Search.Client;
-using NuGetGallery.Configuration;
 
 namespace NuGetGallery
 {
-    public interface IPackageVersionsQuery
-    {
-        Task<IEnumerable<string>> Execute(
-            string id,
-            bool? includePrerelease = false);
-    }
-
-    public class AutocompleteServicePackageVersionsQuery : IPackageVersionsQuery
-    {
-        private readonly ServiceDiscoveryClient _serviceDiscoveryClient;
-        private readonly string _autocompleteServiceResourceType;
-        private readonly RetryingHttpClientWrapper _httpClient;
-
-        public AutocompleteServicePackageVersionsQuery(IAppConfiguration configuration)
-        {
-            _serviceDiscoveryClient = new ServiceDiscoveryClient(configuration.ServiceDiscoveryUri);
-            _autocompleteServiceResourceType = configuration.AutocompleteServiceResourceType;
-            _httpClient = new RetryingHttpClientWrapper(new HttpClient());
-        }
-
-        public async Task<IEnumerable<string>> Execute(string id, bool? includePrerelease)
-        {
-            if (string.IsNullOrWhiteSpace(id))
-            {
-                throw new ArgumentNullException(nameof(id));
-            }
-
-            var queryString = "id=" + Uri.EscapeUriString(id);
-            if (!includePrerelease.HasValue)
-            {
-                queryString += "&prerelease=false";
-            }
-            else
-            {
-                queryString += "&prerelease=" + includePrerelease.Value;
-            }
-
-            var endpoints = await _serviceDiscoveryClient.GetEndpointsForResourceType(_autocompleteServiceResourceType);
-            endpoints = endpoints.Select(e => new Uri(e + "?" + queryString)).AsEnumerable();
-
-            var result = await _httpClient.GetStringAsync(endpoints);
-            var resultObject = JObject.Parse(result);
-
-            return resultObject["data"].Select(entry => entry.ToString());
-        }
-    }
-
     public class PackageVersionsQuery : IPackageVersionsQuery
     {
         private const string SqlFormat = @"SELECT p.[Version]
